@@ -1,28 +1,33 @@
 <?php
 
 /**
- * Class xudfLogEntry
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
  *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+declare(strict_types=1);
+
 class xudfLogEntry extends ActiveRecord
 {
+    public const TABLE_NAME = 'xudf_log_entry';
 
-    const TABLE_NAME = 'xudf_log_entry';
-
-
-    /**
-     * @return string
-     */
-    public function getConnectorContainerName()
+    public function getConnectorContainerName(): string
     {
         return self::TABLE_NAME;
     }
 
-
     /**
-     * @var int
-     *
      * @con_has_field    true
      * @con_sequence     true
      * @con_fieldtype    integer
@@ -30,53 +35,43 @@ class xudfLogEntry extends ActiveRecord
      * @con_is_notnull   true
      * @con_is_primary   true
      */
-    protected $id;
+    protected ?int $id;
+
     /**
-     * @var int
-     *
      * @con_has_field    true
      * @con_fieldtype    integer
      * @con_length       8
      * @con_is_notnull   true
      */
-    protected $obj_id;
+    protected int $obj_id;
+
     /**
-     * @var int
-     *
      * @con_has_field    true
      * @con_fieldtype    integer
      * @con_length       8
      * @con_is_notnull   true
      */
-    protected $usr_id;
+    protected int $usr_id;
+
     /**
-     * @var array
-     *
      * @con_has_field  true
      * @con_fieldtype  clob
      * @con_is_notnull true
      */
-    protected $values = [];
+    protected array $values = [];
+
     /**
-     * @var ilDateTime
-     *
      * @con_has_field  true
      * @con_fieldtype  timestamp
      * @con_index      true
      * @con_is_notnull true
      */
-    protected $timestamp;
-
+    protected ilDateTime $timestamp;
 
     /**
-     * @param $obj_id int
-     * @param $usr_id int
-     * @param $values array
-     *
-     * @return xudfLogEntry
      * @throws ilDateTimeException
      */
-    public static function createNew(int $obj_id, int $usr_id, array $values) : self
+    public static function createNew(int $obj_id, int $usr_id, array $values): self
     {
         $new = new self();
         $new->obj_id = $obj_id;
@@ -88,62 +83,39 @@ class xudfLogEntry extends ActiveRecord
         return $new;
     }
 
-
-    /**
-     * @return int
-     */
-    public function getId() : int
+    public function getId(): int
     {
         return $this->id;
     }
 
-
-    /**
-     * @return int
-     */
-    public function getObjId() : int
+    public function getObjId(): int
     {
         return $this->obj_id;
     }
 
-
-    /**
-     * @return int
-     */
-    public function getUsrId() : int
+    public function getUsrId(): int
     {
         return $this->usr_id;
     }
 
-
-    /**
-     * @return array
-     */
-    public function getValues() : array
+    public function getValues(): array
     {
         return $this->values;
     }
 
-
-    /**
-     * @return ilDateTime
-     */
-    public function getTimestamp() : ilDateTime
+    public function getTimestamp(): ilDateTime
     {
         return $this->timestamp;
     }
 
-
     /**
-     * @param $field_name
-     *
      * @return false|int|mixed|string
      */
-    public function sleep($field_name)
+    public function sleep($field_name): mixed
     {
         switch ($field_name) {
             case 'values':
-                return json_encode($this->values);
+                return json_encode($this->values, JSON_THROW_ON_ERROR);
             case 'timestamp':
                 return $this->timestamp->get(IL_CAL_DATETIME);
             default:
@@ -151,19 +123,15 @@ class xudfLogEntry extends ActiveRecord
         }
     }
 
-
     /**
-     * @param $field_name
-     * @param $field_value
-     *
      * @return ilDateTime|mixed
      * @throws ilDateTimeException
      */
-    public function wakeUp($field_name, $field_value)
+    public function wakeUp($field_name, $field_value): mixed
     {
         switch ($field_name) {
             case 'values':
-                return json_decode($field_value, true);
+                return json_decode($field_value, true, 512, JSON_THROW_ON_ERROR);
             case 'timestamp':
                 return new ilDateTime($field_value, IL_CAL_DATETIME);
         }
