@@ -1,76 +1,70 @@
 <?php
 
-use srag\DIC\UdfEditor\DICTrait;
-use srag\DIC\UdfEditor\Exception\DICException;
-
 /**
- * Class xudfFormConfigurationTableGUI
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
  *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+declare(strict_types=1);
+
+use ILIAS\DI\Container;
+
 class xudfFormConfigurationTableGUI extends ilTable2GUI
 {
+    public const PLUGIN_CLASS_NAME = ilUdfEditorPlugin::class;
 
-    use DICTrait;
-    const PLUGIN_CLASS_NAME = ilUdfEditorPlugin::class;
-    /**
-     * @var ilUdfEditorPlugin
-     */
-    protected $pl;
-
+    protected ilUdfEditorPlugin $pl;
+    private Container $dic;
 
     /**
-     * xudfFormConfigurationTableGUI constructor.
-     *
-     * @param $parent_gui
-     * @param $parent_cmd
-     *
-     * @throws DICException
-     * @throws arException
+     * @throws arException|ilCtrlException
      */
-    public function __construct($parent_gui, $parent_cmd)
+    public function __construct(object $parent_gui, string $parent_cmd)
     {
+        global $DIC;
+        $this->dic = $DIC;
         $this->pl = ilUdfEditorPlugin::getInstance();
 
         parent::__construct($parent_gui, $parent_cmd);
 
-        $this->setFormAction(self::dic()->ctrl()->getFormAction($parent_gui));
-        $this->setRowTemplate(self::plugin()->directory() . '/templates/default/tpl.form_configuration_table_row.html');
+        $this->setFormAction($this->dic->ctrl()->getFormAction($parent_gui));
+        $this->setRowTemplate($this->pl->getDirectory() . '/templates/default/tpl.form_configuration_table_row.html');
 
-        self::dic()->ui()->mainTemplate()->addJavaScript(self::plugin()->directory() . '/templates/default/sortable.js');
-        self::dic()->ui()->mainTemplate()->addJavaScript(self::plugin()->directory() . '/templates/default/waiter.js');
-        self::dic()->ui()->mainTemplate()->addCss(self::plugin()->directory() . '/templates/default/waiter.css');
-        self::dic()->ui()->mainTemplate()->addOnLoadCode("xoctWaiter.init();");
+        $this->dic->ui()->mainTemplate()->addJavaScript($this->pl->getDirectory() . '/templates/default/sortable.js');
+        $this->dic->ui()->mainTemplate()->addJavaScript($this->pl->getDirectory() . '/templates/default/waiter.js');
+        $this->dic->ui()->mainTemplate()->addCss($this->pl->getDirectory() . '/templates/default/waiter.css');
+        $this->dic->ui()->mainTemplate()->addOnLoadCode("xoctWaiter.init();");
 
-        $base_link = self::dic()->ctrl()->getLinkTarget($parent_gui, xudfFormConfigurationGUI::CMD_REORDER, '', true);
-        self::dic()->ui()->mainTemplate()->addOnLoadCode("xudf = {'base_link': '$base_link'};");
+        $base_link = $this->dic->ctrl()->getLinkTarget($parent_gui, xudfFormConfigurationGUI::CMD_REORDER, '', true);
+        $this->dic->ui()->mainTemplate()->addOnLoadCode("xudf = {'base_link': '$base_link'};");
 
         $this->initColumns();
         $this->setData(xudfContentElement::where(['obj_id' => ilObjUdfEditor::_lookupObjectId(filter_input(INPUT_GET, 'ref_id'))])->orderBy('sort')->getArray());
     }
 
-
-    /**
-     * @throws DICException
-     */
-    protected function initColumns()
+    protected function initColumns(): void
     {
         $this->addColumn('', '', 10, true);
-        $this->addColumn(self::dic()->language()->txt('title'), 'title', 50);
-        $this->addColumn(self::dic()->language()->txt('description'), 'description', 100);
-        $this->addColumn(self::dic()->language()->txt('type'), 'type', 30);
-        $this->addColumn(self::plugin()->translate('udf_type'), 'udf_type', 30);
-        $this->addColumn(self::plugin()->translate('is_required'), 'is_required', 30);
+        $this->addColumn($this->dic->language()->txt('title'), 'title', 50);
+        $this->addColumn($this->dic->language()->txt('description'), 'description', 100);
+        $this->addColumn($this->dic->language()->txt('type'), 'type', 30);
+        $this->addColumn($this->pl->txt('udf_type'), 'udf_type', 30);
+        $this->addColumn($this->pl->txt('is_required'), 'is_required', 30);
         $this->addColumn('', '', 10, true);
     }
 
-
-    /**
-     * @param array $a_set
-     *
-     * @throws DICException
-     */
-    protected function fillRow($a_set)
+    protected function fillRow(array $a_set): void
     {
         $udf_definition = ilUserDefinedFields::_getInstance()->getDefinition($a_set['udf_field']);
 
@@ -79,24 +73,28 @@ class xudfFormConfigurationTableGUI extends ilTable2GUI
         }
 
         $this->tpl->setVariable('ID', $a_set['id']);
-        $this->tpl->setVariable('TITLE',
+        $this->tpl->setVariable(
+            'TITLE',
             $a_set['is_separator'] ?
                 $a_set['title']
-                : ($udf_definition['field_name'] ?: self::plugin()->translate('field_not_found')));
+                : ($udf_definition['field_name'] ?: $this->pl->txt('field_not_found'))
+        );
         $this->tpl->setVariable('DESCRIPTION', $a_set['description']);
-        $this->tpl->setVariable('TYPE', $a_set['is_separator'] ? 'Separator' : self::plugin()->translate('udf_field'));
+        $this->tpl->setVariable('TYPE', $a_set['is_separator'] ? 'Separator' : $this->pl->txt('udf_field'));
 
-        $this->tpl->setVariable('UDF_TYPE',
+        $this->tpl->setVariable(
+            'UDF_TYPE',
             $a_set['is_separator'] ? '&nbsp'
-                : ($udf_definition['field_type'] ? self::plugin()->translate('udf_field_type_' . $udf_definition['field_type']) : self::plugin()->translate('field_not_found')));
+                : ($udf_definition['field_type'] ? $this->pl->txt('udf_field_type_' . $udf_definition['field_type']) : $this->pl->txt('field_not_found'))
+        );
 
         if ($a_set['is_separator']) {
             $udf_required = '&nbsp';
         } else {
             if ($a_set['is_required'] == 1) {
-                $udf_required = '<img src="./templates/default/images/icon_ok.svg">';
+                $udf_required = '<img style="width: 1rem" src="./templates/default/images/standard/icon_ok.svg">';
             } else {
-                $udf_required = '<img src="./templates/default/images/icon_not_ok.svg">';
+                $udf_required = '<img style="width: 1rem" src="./templates/default/images/standard/icon_not_ok.svg">';
             }
         }
 
@@ -105,33 +103,23 @@ class xudfFormConfigurationTableGUI extends ilTable2GUI
         $this->tpl->setVariable('ACTIONS', $this->buildActions($a_set['id']));
     }
 
-
-    /**
-     * @throws DICException
-     */
-    protected function showMissingUdfMessage()
+    protected function showMissingUdfMessage(): void
     {
         static $already_shown;
         if (!$already_shown) {
-            ilUtil::sendFailure(self::plugin()->translate('msg_missing_udf'), true);
+            $this->tpl->setOnScreenMessage("failure", $this->pl->txt('msg_missing_udf'), true);
             $already_shown = true;
         }
     }
 
-
-    /**
-     * @param $id
-     *
-     * @return string
-     */
-    protected function buildActions($id)
+    protected function buildActions($id): string
     {
         $actions = new ilAdvancedSelectionListGUI();
-        $actions->setListTitle(self::dic()->language()->txt('actions'));
-        self::dic()->ctrl()->setParameter($this->parent_obj, 'element_id', $id);
+        $actions->setListTitle($this->dic->language()->txt('actions'));
+        $this->dic->ctrl()->setParameter($this->parent_obj, 'element_id', $id);
 
-        $actions->addItem(self::dic()->language()->txt('edit'), 'edit', self::dic()->ctrl()->getLinkTarget($this->parent_obj, xudfFormConfigurationGUI::CMD_EDIT));
-        $actions->addItem(self::dic()->language()->txt('delete'), 'delete', self::dic()->ctrl()->getLinkTarget($this->parent_obj, xudfFormConfigurationGUI::CMD_DELETE));
+        $actions->addItem($this->dic->language()->txt('edit'), 'edit', $this->dic->ctrl()->getLinkTarget($this->parent_obj, xudfFormConfigurationGUI::CMD_EDIT));
+        $actions->addItem($this->dic->language()->txt('delete'), 'delete', $this->dic->ctrl()->getLinkTarget($this->parent_obj, xudfFormConfigurationGUI::CMD_DELETE));
 
         return $actions->getHTML();
     }
