@@ -1,17 +1,28 @@
 <?php
 
 /**
- * Class xudfPageObject
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
  *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+declare(strict_types=1);
+
 class xudfPageObject extends ilPageObject
 {
+    public const PARENT_TYPE = 'xudf';
 
-    const PARENT_TYPE = 'xudf';
-
-
-    function getParentType()
+    public function getParentType(): string
     {
         return 'xudf';
     }
