@@ -1,17 +1,26 @@
 <?php
 
 /**
- * Class xudfPageObjectConfig
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
  *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+declare(strict_types=1);
+
 class xudfPageObjectConfig extends ilPageConfig
 {
-
-    /**
-     * Init
-     */
-    function init()
+    public function init(): void
     {
         // config
         $this->setPreventHTMLUnmasking(true);
