@@ -1,32 +1,35 @@
 <?php
 
-use srag\DIC\UdfEditor\Exception\DICException;
+/**
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
+ *
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+declare(strict_types=1);
 
 /**
- * Class xudfLogGUI
- *
- * @author            Theodor Truffer <tt@studer-raimann.ch>
- *
  * @ilCtrl_isCalledBy xudfLogGUI: ilObjUdfEditorGUI
  */
 class xudfLogGUI extends xudfGUI
 {
-
-    /**
-     * @inheritDoc
-     * @throws DICException
-     */
-    protected function index()
+    protected function index(): void
     {
         $table = new xudfLogTableGUI($this, self::CMD_STANDARD);
         $this->tpl->setContent($table->getHTML());
     }
 
-
-    /**
-     * @throws DICException
-     */
-    protected function applyFilter()
+    protected function applyFilter(): void
     {
         $table = new xudfLogTableGUI($this, self::CMD_STANDARD);
         $table->writeFilterToSession();
@@ -34,11 +37,7 @@ class xudfLogGUI extends xudfGUI
         $this->index();
     }
 
-
-    /**
-     * @throws DICException
-     */
-    protected function resetFilter()
+    protected function resetFilter(): void
     {
         $table = new xudfLogTableGUI($this, self::CMD_STANDARD);
         $table->resetFilter();
