@@ -1,142 +1,125 @@
 <?php
 
-use srag\DIC\UdfEditor\DICTrait;
-use srag\DIC\UdfEditor\Exception\DICException;
+/**
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
+ *
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+declare(strict_types=1);
+
+use ILIAS\DI\Container;
 use srag\Plugins\UdfEditor\Exception\UDFNotFoundException;
 
 require_once __DIR__ . "/../vendor/autoload.php";
 
 /**
- * Class ilObjUdfEditorGUI
- *
- * @author            Theodor Truffer <tt@studer-raimann.ch>
- *
  * @ilCtrl_isCalledBy ilObjUdfEditorGUI: ilRepositoryGUI, ilObjPluginDispatchGUI, ilAdministrationGUI
  * @ilCtrl_Calls      ilObjUdfEditorGUI: ilPermissionGUI, ilInfoScreenGUI, ilObjectCopyGUI, ilCommonActionDispatcherGUI, ilEditClipboardGUI
  */
 class ilObjUdfEditorGUI extends ilObjectPluginGUI
 {
+    public const PLUGIN_CLASS_NAME = ilUdfEditorPlugin::class;
+    public const TAB_CONTENT = 'content';
+    public const TAB_INFO = 'info';
+    public const TAB_SETTINGS = 'settings';
+    public const TAB_HISTORY = 'log_history';
+    public const TAB_PERMISSIONS = 'permissions';
+    public const CMD_INDEX = 'index';
+    public const CMD_SETTINGS = 'showSettings';
 
-    use DICTrait;
-    const PLUGIN_CLASS_NAME = ilUdfEditorPlugin::class;
-    const TAB_CONTENT = 'content';
-    const TAB_INFO = 'info';
-    const TAB_SETTINGS = 'settings';
-    const TAB_HISTORY = 'log_history';
-    const TAB_PERMISSIONS = 'permissions';
-    const CMD_INDEX = 'index';
-    const CMD_SETTINGS = 'showSettings';
+    private Container $dic;
 
-
-    /**
-     * ilObjUdfEditorGUI constructor.
-     */
     public function __construct($a_ref_id = 0, $a_id_type = self::REPOSITORY_NODE_ID, $a_parent_node_id = 0)
     {
         parent::__construct($a_ref_id, $a_id_type, $a_parent_node_id);
+        global $DIC;
+        $this->dic = $DIC;
     }
 
-
-    /**
-     *
-     */
-    public function executeCommand()
+    public function executeCommand(): void
     {
-        $next_class = self::dic()->ctrl()->getNextClass();
-        $cmd = self::dic()->ctrl()->getCmd();
+        $next_class = $this->dic->ctrl()->getNextClass();
+        $cmd = $this->dic->ctrl()->getCmd();
         if (!ilObjUdfEditorAccess::hasReadAccess() && $next_class != strtolower(ilInfoScreenGUI::class) && $cmd != "infoScreen") {
-            ilUtil::sendFailure(self::plugin()->translate('access_denied'), true);
-            self::dic()->ctrl()->returnToParent($this);
+            $this->tpl->setOnScreenMessage("failure", $this->plugin->txt('access_denied'), true);
+            $this->dic->ctrl()->returnToParent($this);
         }
-        if (self::version()->is6()) {
-            $this->tpl->loadStandardTemplate();
-        } else {
-        $this->tpl->getStandardTemplate();
-        }
+        $this->tpl->loadStandardTemplate();
 
         try {
             switch ($next_class) {
                 case strtolower(xudfContentGUI::class):
-                    if (!self::dic()->ctrl()->isAsynch()) {
+                    if (!$this->dic->ctrl()->isAsynch()) {
                         $this->initHeader();
                         $this->setTabs();
                     }
-                    self::dic()->tabs()->activateTab(self::TAB_CONTENT);
+                    $this->dic->tabs()->activateTab(self::TAB_CONTENT);
                     $xvmpGUI = new xudfContentGUI($this);
-                    self::dic()->ctrl()->forwardCommand($xvmpGUI);
-                    if (self::version()->is6()) {
-                        $this->tpl->printToStdout();
-                    } else {
-                    $this->tpl->show();
-                    }
+                    $this->dic->ctrl()->forwardCommand($xvmpGUI);
+                    $this->tpl->printToStdout();
                     break;
                 case strtolower(xudfSettingsGUI::class):
                     if (!ilObjUdfEditorAccess::hasWriteAccess()) {
-                        ilUtil::sendFailure(self::plugin()->translate('access_denied'), true);
-                        self::dic()->ctrl()->returnToParent($this);
+                        $this->tpl->setOnScreenMessage("failure", $this->plugin->txt('access_denied'), true);
+                        $this->dic->ctrl()->returnToParent($this);
                     }
-                    if (!self::dic()->ctrl()->isAsynch()) {
+                    if (!$this->dic->ctrl()->isAsynch()) {
                         $this->initHeader();
                         $this->setTabs();
                     }
-                    self::dic()->tabs()->activateTab(self::TAB_SETTINGS);
+                    $this->dic->tabs()->activateTab(self::TAB_SETTINGS);
                     $xvmpGUI = new xudfSettingsGUI($this);
-                    self::dic()->ctrl()->forwardCommand($xvmpGUI);
-                    if (self::version()->is6()) {
-                        $this->tpl->printToStdout();
-                    } else {
-                    $this->tpl->show();
-                    }
+                    $this->dic->ctrl()->forwardCommand($xvmpGUI);
+                    $this->tpl->printToStdout();
                     break;
                 case strtolower(xudfFormConfigurationGUI::class):
                     if (!ilObjUdfEditorAccess::hasWriteAccess()) {
-                        ilUtil::sendFailure(self::plugin()->translate('access_denied'), true);
-                        self::dic()->ctrl()->returnToParent($this);
+                        $this->tpl->setOnScreenMessage("failure", $this->plugin->txt('access_denied'), true);
+                        $this->dic->ctrl()->returnToParent($this);
                     }
-                    if (!self::dic()->ctrl()->isAsynch()) {
+                    if (!$this->dic->ctrl()->isAsynch()) {
                         $this->initHeader();
                         $this->setTabs();
                     }
-                    self::dic()->tabs()->activateTab(self::TAB_SETTINGS);
+                    $this->dic->tabs()->activateTab(self::TAB_SETTINGS);
                     $xvmpGUI = new xudfFormConfigurationGUI($this);
-                    self::dic()->ctrl()->forwardCommand($xvmpGUI);
-                    if (self::version()->is6()) {
-                        $this->tpl->printToStdout();
-                    } else {
-                    $this->tpl->show();
-                    }
+                    $this->dic->ctrl()->forwardCommand($xvmpGUI);
+                    $this->tpl->printToStdout();
                     break;
                 case strtolower(xudfLogGUI::class):
                     if (!ilObjUdfEditorAccess::hasWriteAccess()) {
-                        ilUtil::sendFailure(self::plugin()->translate('access_denied'), true);
-                        self::dic()->ctrl()->returnToParent($this);
+                        $this->tpl->setOnScreenMessage("failure", $this->plugin->txt('access_denied'), true);
+                        $this->dic->ctrl()->returnToParent($this);
                     }
-                    if (!self::dic()->ctrl()->isAsynch()) {
+                    if (!$this->dic->ctrl()->isAsynch()) {
                         $this->initHeader();
                         $this->setTabs();
                     }
-                    self::dic()->tabs()->activateTab(self::TAB_HISTORY);
+                    $this->dic->tabs()->activateTab(self::TAB_HISTORY);
                     $xvmpGUI = new xudfLogGUI($this);
-                    self::dic()->ctrl()->forwardCommand($xvmpGUI);
-                    if (self::version()->is6()) {
-                        $this->tpl->printToStdout();
-                    } else {
-                    $this->tpl->show();
-                    }
+                    $this->dic->ctrl()->forwardCommand($xvmpGUI);
+                    $this->tpl->printToStdout();
                     break;
                 case strtolower(ilInfoScreenGUI::class):
-                    if (!self::dic()->ctrl()->isAsynch()) {
+                    if (!$this->dic->ctrl()->isAsynch()) {
                         $this->initHeader();
                         $this->setTabs();
                     }
-                    self::dic()->tabs()->activateTab(self::TAB_INFO);
+                    $this->dic->tabs()->activateTab(self::TAB_INFO);
                     $this->checkPermission("visible");
                     $this->infoScreen();    // forwards command
-                    if (self::version()->is6()) {
-                        $this->tpl->printToStdout();
-                    } else {
-                    $this->tpl->show();
-                    }
+                    $this->tpl->printToStdout();
                     break;
                 case strtolower(ilPermissionGUI::class):
                     $this->initHeader(false);
@@ -152,67 +135,34 @@ class ilObjUdfEditorGUI extends ilObjectPluginGUI
                     break;
             }
         } catch (Exception $e) {
-            ilUtil::sendFailure($e->getMessage());
+            $this->tpl->setOnScreenMessage("failure", $e->getMessage());
             if (!$this->creation_mode) {
-                if (self::version()->is6()) {
-                    $this->tpl->printToStdout();
-                } else {
-                $this->tpl->show();
-                }
+                $this->tpl->printToStdout();
             }
         }
     }
 
-
-    /**
-     * @return int
-     */
-    public function getObjId()
+    public function getObjId(): int
     {
         return $this->obj_id;
     }
 
-
-    /**
-     * @return ilObjUdfEditor
-     */
-    public function getObject()
-    {
-        return $this->object;
-    }
-
-
-    /**
-     * @param $cmd
-     */
-    protected function performCommand($cmd)
+    public function performCommand(string $cmd): void
     {
         $this->{$cmd}();
     }
 
-
-    /**
-     *
-     */
-    protected function index()
+    protected function index(): void
     {
-        self::dic()->ctrl()->redirectByClass(xudfContentGUI::class);
+        $this->dic->ctrl()->redirectByClass(xudfContentGUI::class);
     }
 
-
-    /**
-     *
-     */
-    protected function showSettings()
+    protected function showSettings(): void
     {
-        self::dic()->ctrl()->redirectByClass(xudfSettingsGUI::class);
+        $this->dic->ctrl()->redirectByClass(xudfSettingsGUI::class);
     }
 
-
-    /**
-     *
-     */
-    protected function initHeader($render_locator = true)
+    protected function initHeader($render_locator = true): void
     {
         if ($render_locator) {
             $this->setLocator();
@@ -230,93 +180,65 @@ class ilObjUdfEditorGUI extends ilObjectPluginGUI
         }
     }
 
-
-    /**
-     * @return bool
-     */
-    protected function setTabs()
+    protected function setTabs(): void
     {
         global $DIC;
-        $lng = $DIC['lng'];
+        $lng = $DIC->language();
 
-        self::dic()->tabs()->addTab(self::TAB_CONTENT, self::dic()->language()->txt(self::TAB_CONTENT), self::dic()->ctrl()->getLinkTargetByClass(xudfContentGUI::class, xudfContentGUI::CMD_STANDARD));
+        $this->dic->tabs()->addTab(self::TAB_CONTENT, $this->dic->language()->txt(self::TAB_CONTENT), $this->dic->ctrl()->getLinkTargetByClass(xudfContentGUI::class, xudfContentGUI::CMD_STANDARD));
 
         if (xudfSetting::find($this->obj_id)->isShowInfoTab()) {
-            self::dic()->tabs()->addTab(self::TAB_INFO, self::dic()->language()->txt(self::TAB_INFO . '_short'), self::dic()->ctrl()->getLinkTargetByClass(ilInfoScreenGUI::class));
+            $this->dic->tabs()->addTab(self::TAB_INFO, $this->dic->language()->txt(self::TAB_INFO . '_short'), $this->dic->ctrl()->getLinkTargetByClass(ilInfoScreenGUI::class));
         }
 
         if (ilObjUdfEditorAccess::hasWriteAccess()) {
-            self::dic()
+            $this->dic
                 ->tabs()
-                ->addTab(self::TAB_SETTINGS, self::dic()->language()->txt(self::TAB_SETTINGS), self::dic()->ctrl()->getLinkTargetByClass(xudfSettingsGUI::class, xudfSettingsGUI::CMD_STANDARD));
+                ->addTab(self::TAB_SETTINGS, $this->dic->language()->txt(self::TAB_SETTINGS), $this->dic->ctrl()->getLinkTargetByClass(xudfSettingsGUI::class, xudfSettingsGUI::CMD_STANDARD));
 
-            self::dic()->tabs()->addTab(self::TAB_HISTORY, self::dic()->language()->txt('history'), self::dic()->ctrl()->getLinkTargetByClass(xudfLogGUI::class, xudfLogGUI::CMD_STANDARD));
+            $this->dic->tabs()->addTab(self::TAB_HISTORY, $this->dic->language()->txt('history'), $this->dic->ctrl()->getLinkTargetByClass(xudfLogGUI::class, xudfLogGUI::CMD_STANDARD));
         }
 
         if ($this->checkPermissionBool("edit_permission")) {
-            self::dic()->tabs()->addTab("perm_settings", $lng->txt("perm_settings"), self::dic()->ctrl()->getLinkTargetByClass(array(
+            $this->dic->tabs()->addTab("perm_settings", $lng->txt("perm_settings"), $this->dic->ctrl()->getLinkTargetByClass([
                 get_class($this),
                 "ilpermissiongui",
-            ), "perm"));
+            ], "perm"));
         }
-
-        return true;
     }
 
-
-    /**
-     * @param ilInfoScreenGUI $info
-     *
-     * @throws DICException
-     */
-    function addInfoItems($info)
+    public function addInfoItems(ilInfoScreenGUI $info): void
     {
-        $info->addSection(self::plugin()->translate('info_section_title'));
+        $info->addSection($this->plugin->txt('info_section_title'));
         $fields_string = '';
-        foreach (xudfContentElement::where(array('obj_id' => $this->getObjId(), 'is_separator' => 0))->get() as $element) {
+        foreach (xudfContentElement::where(['obj_id' => $this->getObjId(), 'is_separator' => 0])->get() as $element) {
             /** @var $element xudfContentElement */
             try {
                 $fields_string .= $element->getTitle() . '<br>';
             } catch (UDFNotFoundException $e) {
-                self::dic()->logger()->root()->alert($e->getMessage());
-                self::dic()->logger()->root()->alert($e->getTraceAsString());
+                $this->dic->logger()->root()->alert($e->getMessage());
+                $this->dic->logger()->root()->alert($e->getTraceAsString());
             }
         }
-        $info->addProperty(self::plugin()->translate('info_section_subtitle'), $fields_string ? $fields_string : '-');
+        $info->addProperty($this->plugin->txt('info_section_subtitle'), $fields_string ? $fields_string : '-');
     }
 
-
-    /**
-     * @return string
-     */
-    function getAfterCreationCmd()
+    public function getAfterCreationCmd(): string
     {
         return self::CMD_SETTINGS;
     }
 
-
-    /**
-     * @return string
-     */
-    function getStandardCmd()
+    public function getStandardCmd(): string
     {
         return self::CMD_INDEX;
     }
 
-
-    /**
-     * @return string
-     */
-    function getType()
+    public function getType(): string
     {
         return ilUdfEditorPlugin::PLUGIN_ID;
     }
 
-
-    /**
-     * @return bool
-     */
-    protected function supportsCloning()
+    protected function supportsCloning(): bool
     {
         return false;
     }
