@@ -1,8 +1,7 @@
 <?php
-require_once __DIR__ . "/vendor/srag/dic/src/PHPVersionChecker.php";
 $id = "xudf";
 $version = "2.0.0";
-$ilias_min_version = "6.0";
-$ilias_max_version = "7.999";
-$responsible = "fluxlabs ag";
-$responsible_mail = "support@fluxlabs.ch";
+$ilias_min_version = '9.0';
+$ilias_max_version = '9.999';
+$responsible = "mbeym | mjansen";
+$responsible_mail = "mbeym@databay.de mjansen@databay.de";
